commit after changing the lessonname and added field Tags in Assignments
