+++ src/main/java/com/omniwyse/sms/models/Lessons.java
@@ -10,8 +10,8 @@ import javax.persistence.Table;
 public class Lessons {
 	
 	private long id;
+	private String lessondescription;
 	private String lessonname;
-	private String tags;
 	private long subjectid;
 	private long classroomid;
 	private Date lessonstartdate;
@@ -38,17 +38,17 @@ public class Lessons {
 	public void setLessonstartdate(Date lessonstartdate) {
 		this.lessonstartdate = lessonstartdate;
 	}
-	public String getTags() {
-		return tags;
-	}
-	public void setTags(String tags) {
-		this.tags = tags;
-	}
 	public String getLessonname() {
 		return lessonname;
 	}
 	public void setLessonname(String lessonname) {
 		this.lessonname = lessonname;
+	}
+	public String getLessondescription() {
+		return lessondescription;
+	}
+	public void setLessondescription(String lessondescription) {
+		this.lessondescription = lessondescription;
 	}
 	public long getSubjectid() {
 		return subjectid;
