+++ src/main/resources/db/migration/tenants/common/V3__t2_timeline.sql
@@ -2,10 +2,10 @@
 
 CREATE TABLE lessons(
 id bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
-lessonname VARCHAR(150) NOT NULL,
+lessondescription VARCHAR(150) NOT NULL,
 lessonstartdate date NOT NULL,
 status VARCHAR(30),
-tags VARCHAR(50) NOT NULL,
+lessonname VARCHAR(50) NOT NULL,
 subjectid bigint NOT NULL,
 classroomid bigint NOT NULL,
 FOREIGN KEY(classroomid) REFERENCES classrooms(id),
@@ -15,6 +15,7 @@ FOREIGN KEY(subjectid) REFERENCES grade_subjects(id)
 CREATE TABLE assignments(
 id bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
 assignmentname VARCHAR(50) NOT NULL,
+tags VARCHAR(50) NOT NULL,
 dateofassigned date,
 assignmentduedate date,
 classroomid bigint NOT NULL,
